refactor: move UI logic into new module
